namenode/requirements_namenode.txt
Flask
grpcio
grpcio-tools